TVG-102 Blueprint for ShowEpisode endpoints
